working on csv file writer function

+++ app/src/main/java/com/carberry/chris/ece498_mp1/MainActivity.java
@@ -16,6 +16,10 @@ import android.widget.SeekBar;
 
 import java.io.FileWriter;
 import java.io.IOException;
+import java.io.File;
+
+
+import java.io.IOException;
 
 import static android.os.SystemClock.uptimeMillis;
 
@@ -84,10 +88,15 @@ public class MainActivity extends ActionBarActivity {
         currentY = 0;
         numSteps = 0;
 
+
+        //FileWriter writer = new FileWriter("sensor_data.csv");
+
+
+
         // This is what we found on the stack overflow, something is screwed up though, context might be incorrect
         //http://stackoverflow.com/questions/4343342/is-there-a-way-to-retrieve-multiple-sensor-data-in-android
 
-        timeStamp = uptimeMillis();     //time since system boot
+        timeStamp = System.currentTimeMillis();     //time since system boot
 
 
         //private inner class
@@ -115,16 +124,79 @@ public class MainActivity extends ActionBarActivity {
                     // store previous y
                     previousY = currentY;
 
-                } if (sensor.getType() == Sensor.TYPE_GYROSCOPE) {
+                }
+                if (sensor.getType() == Sensor.TYPE_GYROSCOPE) {
                     Gyro_x = event.values[0];
                     Gyro_y = event.values[1];
                     Gyro_z = event.values[2];
-                } if (sensor.getType() == Sensor.TYPE_MAGNETIC_FIELD) {
+                }
+                if (sensor.getType() == Sensor.TYPE_MAGNETIC_FIELD) {
                     Mag_x = event.values[0];
                     Mag_y = event.values[1];
                     Mag_z = event.values[2];
-                } if (sensor.getType() == Sensor.TYPE_LIGHT) {
+                }
+                if (sensor.getType() == Sensor.TYPE_LIGHT) {
                     Light_intensity = event.values[0];
+                }
+                for (int i = 0; i % 1 == 0; i++)
+                {
+                    if (!stopFlag)
+                    {
+                      //write values to
+                      boolean alreadyExists = new File("sensor_data").exists();
+
+                      try
+                      {
+                            if (!alreadyExists)
+                            {
+                                FileWriter writer = new FileWriter("sensor_data.csv");
+                            }
+
+                            boolean i = true;
+                            // Need to fix this while loop with a button?
+
+                            float timeStamp_new = System.currentTimeMillis() - timeStamp;
+
+
+                            writer.append(Float.toString(timeStamp_new));
+                            writer.append(',');
+                            writer.append(Float.toString(Accel_x));
+                            writer.append(',');
+                            writer.append(Float.toString(Accel_y));
+                            writer.append(',');
+                            writer.append(Float.toString(Accel_z));
+                            writer.append(',');
+                            writer.append(Float.toString(Gyro_x));
+                            writer.append(',');
+                            writer.append(Float.toString(Gyro_y));
+                            writer.append(',');
+                            writer.append(Float.toString(Gyro_z));
+                            writer.append(',');
+                            writer.append(Float.toString(Mag_x));
+                            writer.append(',');
+                            writer.append(Float.toString(Mag_y));
+                            writer.append(',');
+                            writer.append(Float.toString(Mag_z));
+                            writer.append(',');
+                            writer.append(Float.toString(Light_intensity));
+                            writer.append('\n');
+
+                            //generate whatever data you want
+
+                            writer.flush();
+                            writer.close();
+                       }
+                       catch(IOException e)
+                       {
+                            e.printStackTrace();
+                       }
+
+                    }
+
+                    else {
+
+                        }
+                    }
                 }
             }
 
@@ -190,13 +262,23 @@ public class MainActivity extends ActionBarActivity {
     */
     public void generateCsvFile(String sFilename)
     {           //call this function onClick (start)
+
+        boolean alreadyExists = new File(sFilename).exists();
+
         try
         {
-            FileWriter writer = new FileWriter(sFilename);
+            if (!alreadyExists)
+            {
+                FileWriter writer = new FileWriter(sFilename);
+            }
+
             boolean i = true;
             // Need to fix this while loop with a button?
-            while(i) { //while !onClick(Stop)    race condition?
-                writer.append(Float.toString(timeStamp));
+
+                float timeStamp_new = System.currentTimeMillis() - timeStamp;
+
+
+                writer.append(Float.toString(timeStamp_new));
                 writer.append(',');
                 writer.append(Float.toString(Accel_x));
                 writer.append(',');
@@ -218,8 +300,6 @@ public class MainActivity extends ActionBarActivity {
                 writer.append(',');
                 writer.append(Float.toString(Light_intensity));
                 writer.append('\n');
-
-            }
 
             //generate whatever data you want
